services/alert/requirements.txt
redis==5.0.1
kafka-python==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
alembic==1.13.0
pytest==7.4.3
# Azure Event Hubs
azure-eventhub==5.11.4
# Slack SDK
slack-sdk==3.26.1
